% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findglobalarg.r
\name{find_global_arg}
\alias{find_global_arg}
\title{Find the global arguments not given by the argument list}
\usage{
find_global_arg(fn)
}
\arguments{
\item{fn:}{a function to be checked}
}
\value{
a vector of string
}
\description{
Find the global arguments not given by the argument list
}
\examples{
find_global_arg(fn)
}
\keyword{coding}
