% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_wd.r
\name{get_parent_wd}
\alias{get_parent_wd}
\title{Get the parent working directory of the current file}
\usage{
get_parent_wd()
}
\value{
a string of the parent directory
}
\description{
Get the parent working directory of the current file
}
\examples{
get_parent_wd()
}
\keyword{directory}
