% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_conjugate_update.r
\name{normal_standard_update_mean}
\alias{normal_standard_update_mean}
\title{Update the mean and sd of nromal kernel with normal prior on mean.}
\usage{
normal_standard_update_mean(x_bar, sigma, n, mu_0, sigma_0)
}
\arguments{
\item{x_bar:}{the mean of data sample}

\item{sigma:}{the sd of sampling distribution, should be known}

\item{n:}{number of data points}

\item{mu_0:}{the mean of the normal prior}

\item{sigma_0:}{the sd of the normal prior}
}
\value{
: a random draw from the posterior normal distribution.
}
\description{
Update the mean and sd of nromal kernel with normal prior on mean.
}
\examples{
normal_standard_update(1, 1, n, 1, 1)
}
\keyword{conjugate}
\keyword{update}
