% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_conjugate_update.r
\name{normal_standard_update_var}
\alias{normal_standard_update_var}
\title{Update the sd of nromal kernel with inverse gamma prior}
\usage{
normal_standard_update_var(mu, dat, n, sigma_a, sigma_b, re_sd = T)
}
\arguments{
\item{mu:}{the mean of sampling distribution, known}

\item{dat:}{the observed normal draws}

\item{n:}{number of data points}

\item{sigma_a:}{the IG prior shape}

\item{sigma_b:}{the IG prior rate}

\item{re_sd:}{boulean, true if return sd, o.w. return variance.}
}
\value{
: a random draw from IG, or the sqrt of that, depending on re_sd.
}
\description{
Update the sd of nromal kernel with inverse gamma prior
}
\examples{
normal_standard_update_var(1, 1, n, 1, 1)
}
\keyword{conjugate}
\keyword{update}
