% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.r
\name{plot_acf}
\alias{plot_acf}
\title{plot the acf plot of a mcmc chain of a parameter}
\usage{
plot_acf(param_sample, bquote_param_name)
}
\arguments{
\item{param_sample:}{a mcmc chain of one parameter}

\item{bquote_param_name:}{a bquote object}
}
\value{
: a plot of acf plot
}
\description{
plot the acf plot of a mcmc chain of a parameter
}
\examples{
x = rnorm(1000)
plot_acf(x, 'x')
}
\keyword{acfplot}
