% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.r
\name{plot_posterior_density}
\alias{plot_posterior_density}
\title{Plot the density of raw data, point estimates and confidence interval.}
\usage{
plot_posterior_density(predictive_sample, dat, plot_name,
  plot_ci = FALSE, plot_hist = TRUE, legend_pos = "topright", ...)
}
\arguments{
\item{predictive_sample:}{the point estimates}

\item{dat:}{raw data}

\item{plot_name:}{name of the plot}

\item{plot_ci:}{boolean, default false, when true, check for additonal arguments to plot the confidence interval}

\item{plot_hist:}{boolean, default true, when true, plot the histogram of raw data.}

\item{legend_pos:}{string, default 'topright', can be changed to 'topleft'.}

\item{lb:}{vector of lowerbound values of CI}

\item{ub:}{vector of upperbound values of CI}

\item{grid:}{vector of grid based on which lb and ub are calculated.}

\item{level:}{str, the confidence level.}
}
\description{
Plot the density of raw data, point estimates and confidence interval.
}
\keyword{'Test}
\keyword{2,}
\keyword{=}
\keyword{by=0.01),}
\keyword{density}
\keyword{dnorm(seq(-2,}
\keyword{grid=seq(-2,}
\keyword{lb}
\keyword{lb=lb,}
\keyword{level='0.9')}
\keyword{mean=0,}
\keyword{plot}
\keyword{plot',}
\keyword{plot_ci=T,}
\keyword{plot_posterior_density(sim_data,}
\keyword{rnorm(1000)}
\keyword{sd=0.9)}
\keyword{sd=1.1)}
\keyword{sim_data}
\keyword{sim_data+0.01,}
\keyword{ub}
\keyword{ub=ub,}
