% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_hist_ci_mean.r
\name{plot_posterior_hist_ci_mean}
\alias{plot_posterior_hist_ci_mean}
\title{Plot the histogram of posterior sample with verticle blue lines indicating the 95% confidence interval and median, red line indicating the true value.}
\usage{
plot_posterior_hist_ci_mean(param_sample, bquote_param_name,
  true_value = NULL)
}
\arguments{
\item{param_sample:}{a mcmc chain of one parameter}

\item{bquote_param_name:}{a bquote object}

\item{true_value:}{the true value of parameter}
}
\value{
a plot of histogram
}
\description{
Plot the histogram of posterior sample with verticle blue lines indicating the 95% confidence interval and median, red line indicating the true value.
}
\examples{
x = rnorm(1000)
plot_posterior_hist_ci_mean(x, 'x', 0)
}
\keyword{histogram}
\keyword{posterior}
