% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traceplot.r
\name{plot_traceplot}
\alias{plot_traceplot}
\title{plot the trace plot of a mcmc chain of a parameter}
\usage{
plot_traceplot(param_sample, bquote_param_name, sample_mean = F,
  sample_median = F, plot_value = F, ind_value = NULL)
}
\arguments{
\item{param_sample:}{a mcmc chain of one parameter}

\item{bquote_param_name:}{a bquote object}

\item{sample_mean:}{boolean, if plot sample mean}

\item{sample_median:}{boolean, if plot sample median}

\item{plot_value:}{boolean, if plot the ind_value}

\item{ind_value:}{a float, the special value}
}
\value{
plot of traceplot
}
\description{
plot the trace plot of a mcmc chain of a parameter
}
\examples{
x = rnorm(1000)
plot_traceplot(x, 'x', sample_mean=T, sample_median=T, plot_value=T, ind_value=-1)
}
\keyword{traceplot}
