% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_samples.r
\name{sum_samples}
\alias{sum_samples}
\title{Summarize the mcmc sample that lives in a dataframe, return the summary of each column as a row of present table.}
\usage{
sum_samples(sample_df)
}
\arguments{
\item{sample_df:}{the dataframe whose columns are the mcmc chains of parameters.}
}
\value{
a dataframe that summarizes the posterior samples
}
\description{
Summarize the mcmc sample that lives in a dataframe, return the summary of each column as a row of present table.
}
\examples{
x = rnorm(1000)
y = rnorm(1000, 10, 1)
output=cbind(x=x, y=y)
sum_samples(output)
}
\keyword{posterior}
\keyword{summary}
